requirements.txt
streamlit
pandas
numpy
python-dotenv
tushare
google-generativeai
dashscope
openai
pdfkit
requests